+++ CHANGELOG.md
@@ -5,7 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased (1.3.0)
+## [1.2.6] 
+
+### Added 
+
+- `owner()` view function to retrieve contract owner address
+
+### Changed 
+
+- `uint8 status` parameter instead of `int status` in the `initialize` function
+
+## [1.2.5] - 2020-09-08
+
+### Changed
+
+- KeystoreProvider dependency updated to support deployments via HTTPS
+
+## [1.2.4] - 2020-08-18
+
+### Added
+
+- KeystoreProvider for deployments using an Ethereum keystore file
 
 ## [1.2.3] - 2020-03-10
 
@@ -82,6 +102,9 @@ Minor modifications after running the [slither] source analyzer
 [1.2.1]: https://github.com/clearmatics/asset-token/compare/v1.2.0...v1.2.1
 [1.2.2]: https://github.com/clearmatics/asset-token/compare/v1.2.1...v1.2.2
 [1.2.3]: https://github.com/clearmatics/asset-token/compare/v1.2.2...v1.2.3
+[1.2.4]: https://github.com/clearmatics/asset-token/compare/v1.2.3...v1.2.4
+[1.2.5]: https://github.com/clearmatics/asset-token/compare/v1.2.4...v1.2.5
+[1.2.6]: https://github.com/clearmatics/asset-token/compare/v1.2.5...v1.2.6
 [erc777]: https://eips.ethereum.org/EIPS/eip-777
 [erc1820]: https://eips.ethereum.org/EIPS/eip-1820
 [erc20]: https://eips.ethereum.org/EIPS/eip-20